+++ Runtime/RendererTintSelectionVisual.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace JorisHoef.ObjectSelection
+namespace Deucarian.ObjectSelection
 {
     /// <summary>
     /// Selection visual that tints a target renderer with a MaterialPropertyBlock.