+++ Runtime/UnityObjectUtility.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace JorisHoef.ObjectSelection
+namespace Deucarian.ObjectSelection
 {
     internal static class UnityObjectUtility
     {